build: fix darwin cross-compilation, remove linux-only binaries from darwin targets
- Makefile: remove dokid, doki-compose, doki-init from darwin targets
  (these require Linux namespaces, cgroups, syscall.Setxattr, CLONE_*)
- pkg/storage/driver.go: cast stat.Bsize to int64 for darwin uint32 type

+++ pkg/storage/driver.go
@@ -145,7 +145,7 @@ func (m *Manager) Stats() StorageStats {
 		// Statfs_t.Bavail is uint64, Bsize is int64 (signed on Android).
 		// Compute the product in uint64 to avoid signed-overflow UB and
 		// then clamp to math.MaxInt64 for gosec G115.
-		bsize := common.SafeUint64FromInt64(stat.Bsize)
+		bsize := common.SafeUint64FromInt64(int64(stat.Bsize))
 		product := stat.Bavail * bsize
 		stats.FreeSpace = common.SafeInt64FromUint64(product)
 	}
